K.table(tolerance)	R Documentation

Tables of K-factors for Tolerance Intervals Based on Normality

Description

Tabulated summary of k-factors for tolerance intervals based on normality. The user can specify multiple values for each of the three inputs.

Usage


K.table(n, alpha, P, side = 1, f = NULL,
        by.arg = c("n", "alpha", "P")) 

Arguments

n 	 A vector of (effective) sample sizes.

alpha 	 The level chosen such that 1-alpha is the confidence level. Can be a vector.

P 	 The proportion of the population to be covered by this tolerance interval. Can be a vector.

side 	 Whether a 1-sided or 2-sided tolerance interval is required (determined by side = 1 or side = 2, respectively).

f 	 The number of degrees of freedom associated with calculating the estimate of the population standard deviation. If NULL, then f is taken to be n-1. Only a single value can be specified for f.

by.arg 	 How you would like the output organized. If by.arg = "n", then the output provides a list of matrices sorted by the values specified in n. The matrices have rows corresponding to the values specified by 1-alpha and columns corresponding to the values specified by P. If by.arg = "alpha", then the output provides a list of matrices sorted by the values specified in 1-alpha. The matrices have rows corresponding to the values specified by n and columns corresponding to the values specified by P. If by.arg = "P", then the output provides a list of matrices sorted by the values specified in P. The matrices have rows corresponding to the values specified by 1-alpha and columns corresponding to the values specified by n.

Details

The method used for estimating the k-factors is that due to Howe as it is generally viewed as more accurate than the Weissberg-Beatty method.

Value

K.table returns a list with a structure determined by the argument by.arg described above.

References

Howe, W. G. (1969), Two-Sided Tolerance Limits for Normal Populations - Some Improvements, Journal of the American Statistical Association, 64, 610–620.

Weissberg, A. and Beatty, G. (1969), Tables of Tolerance Limit Factors for Normal Distributions, Technometrics, 2, 483–500.

See Also

K.factor

Examples

 
## Tables generated for each value of the sample size.

K.table(n = seq(50, 100, 10), alpha = c(0.01, 0.05, 0.10), 
        P = c(0.90, 0.95, 0.99), by.arg = "n")

## Tables generated for each value of the confidence level.

K.table(n = seq(50, 100, 10), alpha = c(0.01, 0.05, 0.10), 
        P = c(0.90, 0.95, 0.99), by.arg = "alpha")

## Tables generated for each value of the coverage proportion.

K.table(n = seq(50, 100, 10), alpha = c(0.01, 0.05, 0.10), 
        P = c(0.90, 0.95, 0.99), by.arg = "P")
        

[Package tolerance version 0.2.3 Index]