acc.samp(tolerance)	R Documentation

Acceptance Sampling

Description

Provides an upper bound on the number of acceptable rejects or nonconformities in a process. This is similar to a 1-sided upper tolerance bound for a hypergeometric random variable.

Usage


acc.samp(n, N, alpha = 0.05, P = 0.99, AQL = 0.01, RQL = 0.02)

Arguments

n 	 The sample size to be drawn from the inventory.

N 	 The total inventory (or lot) size.

alpha 	 1-alpha is the confidence level for bounding the probability of accepting the inventory.

P 	 The proportion of items in the inventory which are to be accountable.

AQL 	 The acceptable quality level, which is the largest proportion of defects in a process considered acceptable. Note that 0 < AQL < 1.

RQL 	 The rejectable quality level, which is the largest proportion of defects in an independent lot that one is willing to tolerate. Note that AQL < RQL < 1.

Value

acc.samp returns a matrix with the following quantities:

acceptance.limit 	 The number of items in the sample which may be unaccountable, yet still be able to attain the desired confidence level 1-alpha.

lot.size 	 The total inventory (or lot) size N.

confidence 	 The confidence level 1-alpha.

P 	 The proportion of accountable items specified by the user.

AQL 	 The acceptable quality level as specified by the user. If the sampling were to be repeated numerous times as a process, then this quantity specifies the proportion of missing items considered acceptable from the process as a whole. Conditioning on the calculated value for acceptance.limit, the AQL is used to estimate the producer's risk (see prod.risk below).

RQL 	 The rejectable quality level as specified by the user. This is the proportion of individual items in a sample one is willing to tolerate missing. Conditioning on the calculated value for acceptance.limit, the RQL is used to estimate the consumer's risk (see cons.risk below).

sample.size 	 The sample size drawn as specified by n.

prod.risk 	 The producer's risk at the specified AQL. This is the probability of rejecting an audit of a good inventory (also called the Type I error). A good inventory can be rejected if an unfortunate random sample is selected (e.g., most of the missing items happened to be selected for the audit). 1-prod.risk gives the confidence level of this sampling plan for the specified AQL and RQL. If it is lower than the confidence level desired (e.g., because the AQL is too high), then a warning message will be displayed.

cons.risk 	 The consumer's risk at the specified RQL. This is the probability of accepting an audit of a bad inventory (also called the Type II error). A bad inventory can be accepted if a fortunate random sample is selected (e.g., most of the missing items happened to not be selected for the audit).

References

Montgomery, D. C. (2005), Introduction to Statistical Quality Control, Fifth Edition, John Wiley & Sons, Inc.

See Also

Hypergeometric

Examples

 
## A 90%/90% acceptance sampling plan for a sample of 450 
## drawn from a lot size of 960.

acc.samp(n = 450, N = 960, alpha = 0.10, P = 0.90, AQL = 0.07,
         RQL = 0.10)
 

[Package tolerance version 0.3.0 Index]