bonftol.int(tolerance)	R Documentation

Approximate 2-Sided Tolerance Intervals that Control the Tails Using Bonferroni's Inequality

Description

This function allows the user to control what proportion of the population is to be in the tails of the given distribution for a 2-sided tolerance interval. The result is a conservative approximation based on Bonferroni's inequality.

Usage


bonftol.int(fn, P1 = 0.005, P2 = 0.005, alpha = 0.05, ...)

Arguments

fn 	 The function name for the 2-sided tolerance interval to be calculated.

P1 	 The proportion of the population not covered in the lower tail of the distribution.

P2 	 The proportion of the population not covered in the upper tail of the distribution.

alpha 	 The level chosen such that 1-alpha is the confidence level.

... 	 Additional arguments passed to fn, including the data. All arguments that would be specified in fn must also be specified here.

Value

The results for the 2-sided tolerance interval procedure are reported. See the corresponding help file for fn about specific output. Note that the (minimum) proportion of the population to be covered by this interval is 1 - (P1 + P2).

Note

This function can be used with any 2-sided tolerance interval function, including the regression tolerance interval functions.

References

Jensen, W. A. (2009), Approximations of Tolerance Intervals for Normally Distributed Data, Quality and Reliability Engineering International, 25, 571–580.

Patel, J. K. (1986), Tolerance Intervals - A Review, Communications in Statistics - Theory and Methodology, 15, 2719–2762.

Examples

 
## 95%/97% tolerance interval for normally distributed
## data controlling 1% of the data is in the lower tail
## and 2% of the data in the upper tail.

set.seed(100)
x <- rnorm(100, 0, 0.2)
bonftol.int(normtol.int, x = x, P1 = 0.01, P2 = 0.02,
            alpha = 0.05, method = "HE")


[Package tolerance version 0.2.3 Index]