cautol.int(tolerance)	R Documentation

Cauchy Tolerance Intervals

Description

Provides 1-sided or 2-sided tolerance intervals for Cauchy distributed data.

Usage


cautol.int(x, alpha = 0.05, P = 0.99, side = 1)

Arguments

x 	 A vector of data which is Cauchy distributed.

alpha 	 The level chosen such that 1-alpha is the confidence level.

P 	 The proportion of the population to be covered by this tolerance interval.

side 	 Whether a 1-sided or 2-sided tolerance interval is required (determined by side = 1 or side = 2, respectively).

Value

cautol.int returns a data.frame with items:

alpha 	 The specified significance level.

P 	 The proportion of the population covered by this tolerance interval.

1-sided.lower 	 The 1-sided lower tolerance bound. This is given only if side = 1.

1-sided.upper 	 The 1-sided upper tolerance bound. This is given only if side = 1.

2-sided.lower 	 The 2-sided lower tolerance bound. This is given only if side = 2.

2-sided.upper 	 The 2-sided upper tolerance bound. This is given only if side = 2.

References

Bain, L. J. (1978), Statistical Analysis of Reliability and Life-Testing Models, Marcel Dekker, Inc.

See Also

Cauchy

Examples

 
## 95%/90% 2-sided Cauchy tolerance interval for a sample 
## of size 1000. 

set.seed(100)
x <- rcauchy(1000, 100000, 10)
out <- cautol.int(x = x, alpha = 0.05, P = 0.90, side = 2)
out

plottol(out, x, plot.type = "both", x.lab = "Cauchy Data")

[Package tolerance version 0.2.3 Index]