exp2tol.int(tolerance)	R Documentation

2-Parameter Exponential Tolerance Intervals

Description

Provides 1-sided or 2-sided tolerance intervals for data distributed according to a 2-parameter exponential distribution. Data with Type II censoring is permitted.

Usage


exp2tol.int(x, alpha = 0.05, P = 0.99, side = 1,
            method = c("GPU", "DUN"), type.2 = FALSE)

Arguments

x 	 A vector of data which is distributed according to the 2-parameter exponential distribution.

alpha 	 The level chosen such that 1-alpha is the confidence level.

P 	 The proportion of the population to be covered by this tolerance interval.

side 	 Whether a 1-sided or 2-sided tolerance interval is required (determined by side = 1 or side = 2, respectively).

method 	 The method for how the upper tolerance bound is approximated. "GPU" is the Guenther-Patil-Upppuluri method. "DUN" is the Dunsmore method, which was empirically shown to be an improvement for samples greater than or equal to 8. More information on these methods can be found in the "References".

type.2 	 Select TRUE if Type II censoring is present (i.e., the data set is censored at the maximum value present). The default is FALSE.

Value

exp2tol.int returns a data frame with items:

alpha 	 The specified significance level.

P 	 The proportion of the population covered by this tolerance interval.

1-sided.lower 	 The 1-sided lower tolerance bound. This is given only if side = 1.

1-sided.upper 	 The 1-sided upper tolerance bound. This is given only if side = 1.

2-sided.lower 	 The 2-sided lower tolerance bound. This is given only if side = 2.

2-sided.upper 	 The 2-sided upper tolerance bound. This is given only if side = 2.

References

Dunsmore, I. R. (1978), Some Approximations for Tolerance Factors for the Two Parameter Exponential Distribution, Technometrics, 20, 317–318.

Engelhardt, M. and Bain, L. J. (1978), Tolerance Limits and Confidence Limits on Reliability for the Two-Parameter Exponential Distribution, Technometrics, 20, 37–39.

Guenther, W. C., Patil, S. A., and Uppuluri, V. R. R. (1976), One-Sided β-Content Tolerance Factors for the Two Parameter Exponential Distribution, Technometrics, 18, 333–340.

See Also

TwoParExponential

Examples

 
## 95%/90% 1-sided 2-parameter exponential tolerance intervals
## for a sample of size 50. 

set.seed(100)
x <- r2exp(50, 6, shift = 55)
out <- exp2tol.int(x = x, alpha = 0.05, P = 0.90, side = 1,
                   method = "DUN", type.2 = FALSE)
out

plottol(out, x, plot.type = "both", side = "upper", 
        x.lab = "2-Parameter Exponential Data")

[Package tolerance version 0.2.3 Index]