exptol.int(tolerance)	R Documentation

Exponential Tolerance Intervals

Description

Provides 1-sided or 2-sided tolerance intervals for data distributed according to an exponential distribution. Data with Type II censoring is permitted.

Usage


exptol.int(x, alpha = 0.05, P = 0.99, side = 1, type.2 = FALSE)

Arguments

x 	 A vector of data which is distributed according to an exponential distribution.

alpha 	 The level chosen such that 1-alpha is the confidence level.

P 	 The proportion of the population to be covered by this tolerance interval.

side 	 Whether a 1-sided or 2-sided tolerance interval is required (determined by side = 1 or side = 2, respectively).

type.2 	 Select TRUE if Type II censoring is present (i.e., the data set is censored at the maximum value present). The default is FALSE.

Value

exptol.int returns a data frame with items:

alpha 	 The specified significance level.

P 	 The proportion of the population covered by this tolerance interval.

lambda.hat 	 The mean of the data (i.e., 1/rate).

1-sided.lower 	 The 1-sided lower tolerance bound. This is given only if side = 1.

1-sided.upper 	 The 1-sided upper tolerance bound. This is given only if side = 1.

2-sided.lower 	 The 2-sided lower tolerance bound. This is given only if side = 2.

2-sided.upper 	 The 2-sided upper tolerance bound. This is given only if side = 2.

References

Blischke, W. R. and Murthy, D. N. P. (2000), Reliability: Modeling, Prediction, and Optimization, John Wiley & Sons, Inc.

See Also

Exponential

Examples

 
## 95%/99% 1-sided exponential tolerance intervals for a
## sample of size 50. 

set.seed(100)
x <- rexp(100, 0.004)
out <- exptol.int(x = x, alpha = 0.05, P = 0.99, side = 1,
                  type.2 = FALSE)
out

plottol(out, x, plot.type = "both", side = "lower", 
        x.lab = "Exponential Data")

[Package tolerance version 0.3.0 Index]