exttol.int(tolerance)	R Documentation

Weibull (or Extreme-Value) Tolerance Intervals

Description

Provides 1-sided or 2-sided tolerance intervals for data distributed according to either a Weibull distribution or extreme-value (also called Gumbel) distributions.

Usage


exttol.int(x, alpha = 0.05, P = 0.99, side = 1,
           dist = c("Weibull", "Gumbel"), ext = c("min", "max"), 
           NR.delta = 1e-8)

Arguments

x 	 A vector of data which is distributed according to either a Weibull distribution or an extreme-value distribution.

alpha 	 The level chosen such that 1-alpha is the confidence level.

P 	 The proportion of the population to be covered by this tolerance interval.

side 	 Whether a 1-sided or 2-sided tolerance interval is required (determined by side = 1 or side = 2, respectively).

dist 	 Select either dist = "Weibull" or dist = "Gumbel" if the data is distributed according to the Weibull or extreme-value distribution, respectively.

ext 	 If dist = "Gumbel", then select which extreme is to be modeled for the Gumbel distribution. The Gumbel distribution for the minimum (i.e., ext = "min") corresponds to a left-skewed distribution and the Gumbel distribution for the maximum (i.e., ext = "max") corresponds to a right-skewed distribution

NR.delta 	 The stopping criterion used for the Newton-Raphson algorithm when finding the maximum likelihood estimates of the Weibull or extreme-value distribution.

Details

Recall that the relationship between the Weibull distribution and the extreme-value distribution for the minimum is that if the random variable X is distributed according to a Weibull distribution, then the random variable Y = ln(X) is distributed according to an extreme-value distribution for the minimum.

If dist = "Weibull", then the natural logarithm of the data are taken so that a Newton-Raphson algorithm can be employed to find the MLEs of the extreme-value distribution for the minimum and then the data and MLEs are transformed back appropriately. No transformation is performed if dist = "Gumbel". The Newton-Raphson algorithm is initialized by the method of moments estimators for the parameters.

Value

exttol.int returns a data frame with items:

alpha 	 The specified significance level.

P 	 The proportion of the population covered by this tolerance interval.

shape.1 	 MLE for the shape parameter if dist = "Weibull" or for the location parameter if dist = "Gumbel".

shape.2 	 MLE for the scale parameter if dist = "Weibull" or dist = "Gumbel".

1-sided.lower 	 The 1-sided lower tolerance bound. This is given only if side = 1.

1-sided.upper 	 The 1-sided upper tolerance bound. This is given only if side = 1.

2-sided.lower 	 The 2-sided lower tolerance bound. This is given only if side = 2.

2-sided.upper 	 The 2-sided upper tolerance bound. This is given only if side = 2.

References

Bain, L. J. and Engelhardt, M. (1981), Simple Approximate Distributional Results for Confidence and Tolerance Limits for the Weibull Distribution Based on Maximum Likelihood Estimators, Technometrics, 23, 15–20.

Coles, S. (2001), An Introduction to Statistical Modeling of Extreme Values, Springer.

See Also

Weibull

Examples

 
## 85%/90% 1-sided Weibull tolerance intervals for a sample
## of size 150. 

set.seed(100)
x <- rweibull(150, 3, 75)
out <- exttol.int(x = x, alpha = 0.15, P = 0.90, side = 1,
                  dist = "Weibull")
out

plottol(out, x, plot.type = "both", side = "lower", 
        x.lab = "Weibull Data")

[Package tolerance version 0.2.3 Index]