logistol.int(tolerance)	R Documentation

Logistic (or Log-Logistic) Tolerance Intervals

Description

Provides 1-sided or 2-sided tolerance intervals for data distributed according to a logistic or log-logistic distribution.

Usage


logistol.int(x, alpha = 0.05, P = 0.99, log.log = FALSE,
             side = 1)

Arguments

x 	 A vector of data which is distributed according to a logistic or log-logistic distribution.

alpha 	 The level chosen such that 1-alpha is the confidence level.

P 	 The proportion of the population to be covered by this tolerance interval.

log.log 	 If TRUE, then the data is considered to be from a log-logistic distribution, in which case the output gives tolerance intervals for the log-logistic distribution. The default is FALSE.

side 	 Whether a 1-sided or 2-sided tolerance interval is required (determined by side = 1 or side = 2, respectively).

Details

Recall that if the random variable X is distributed according to a log-logistic distribution, then the random variable Y = ln(X) is distributed according to a logistic distribution.

Value

logistol.int returns a data frame with items:

alpha 	 The specified significance level.

P 	 The proportion of the population covered by this tolerance interval.

1-sided.lower 	 The 1-sided lower tolerance bound. This is given only if side = 1.

1-sided.upper 	 The 1-sided upper tolerance bound. This is given only if side = 1.

2-sided.lower 	 The 2-sided lower tolerance bound. This is given only if side = 2.

2-sided.upper 	 The 2-sided upper tolerance bound. This is given only if side = 2.

References

Balakrishnan, N. (1992), Handbook of the Logistic Distribution, Marcel Dekker, Inc.

Hall, I. J. (1975), One-Sided Tolerance Limits for a Logistic Distribution Based on Censored Samples, Biometrics, 31, 873–880.

See Also

Logistic

Examples

 
## 90%/95% 1-sided logistic tolerance intervals for a sample
## of size 20. 

set.seed(100)
x <- rlogis(20, 5, 1)
out <- logistol.int(x = x, alpha = 0.10, P = 0.95, 
                    log.log = FALSE, side = 1) 
out

plottol(out, x, plot.type = "control", side = "two", 
        x.lab = "Logistic Data")

[Package tolerance version 0.2.3 Index]