mvregtol.region(tolerance)	R Documentation

Multivariate (Multiple) Linear Regression Tolerance Regions

Description

Determines the appropriate tolerance factor for computing multivariate (multiple) linear regression tolerance regions based on Monte Carlo simulation.

Usage


mvregtol.region(y, x, new.x = NULL, int = TRUE, alpha = 0.05, 
                P = 0.99, B = 1000)

Arguments

y 	 An nxq matrix of responses assumed to be drawn from a q-dimensional multivariate normal distribution. n pertains to the sample size.

x 	 An nxm matrix of predictors (i.e., the design matrix). m pertains to the number of predictors. Do NOT include a column of 1's if assuming an intercept. This is controlled by the int argument.

new.x 	 A matrix of new values for which to approximate k-factors. This must be a matrix with m columns.

int 	 If int = TRUE, then an intercept is assumed.

alpha 	 The level chosen such that 1-alpha is the confidence level.

P 	 The proportion of the population to be covered by this tolerance region.

B 	 The number of iterations used for the Monte Carlo algorithm which determines the tolerance factor. The number of iterations should be at least as large as the default value of 1000.

Details

A basic sketch of how the algorithm works is as follows:

(1) Generate independent chi-square random variables and Wishart random matrices.

(2) Compute the eigenvalues of the randomly generated Wishart matrices.

(3) Iterate the above steps to generate a set of B sample values such that the 100(1-alpha)-th percentile is an approximate tolerance factor.

Value

mvregtol.region returns a matrix where the first column is the k-factor, the next q columns are the estimated responses from the least squares fit, and the final m columns are the predictor values. The first n rows of the matrix pertain to the raw data as specified by y and x. If values for new.x are specified, then there is one additional row appended to this output for each row in the matrix new.x.

References

Anderson, T. W. (2003) An Introduction to Multivariate Statistical Analysis, Third Edition, Wiley.

Krishnamoorthy, K. and Mathew, T. (2009), Statistical Tolerance Regions: Theory, Applications, and Computation, Wiley.

Krishnamoorthy, K. and Mondal, S. (2008), Tolerance Factors in Multiple and Multivariate Linear Regressions, Communications in Statistics - Simulation and Computation, 37, 546–559.

Examples

 
## 95%/95% multivariate regression tolerance factors using
## a fertilizer data set presented in Anderson (2003, p. 374). 

grain <- c(40, 17, 9, 15, 6, 12, 5, 9)
straw <- c(53, 19, 10, 29, 13, 27, 19, 30)
fert <- c(24, 11, 5, 12, 7, 14, 11, 18)
Y <- cbind(grain, straw)
X <- cbind(fert)
new.x <- c(10, 15, 20)

set.seed(100)
out <- mvregtol.region(Y, X, new.x, int = TRUE, alpha = 0.05,
                       P = 0.95, B = 5000)
out

[Package tolerance version 0.2.3 Index]