nptol.int(tolerance)	R Documentation

Nonparametric Tolerance Intervals

Description

Provides 1-sided or 2-sided nonparametric (i.e., distribution-free) tolerance intervals for any continuous data set.

Usage


nptol.int(x, alpha = 0.05, P = 0.99, side = 1,
          method = c("WILKS", "WALD", "HM"), upper = NULL,
          lower = NULL)

Arguments

x 	 A vector of data which no distributional assumptions are made. The data is only assumed to come from a continuous distribution.

alpha 	 The level chosen such that 1-alpha is the confidence level.

P 	 The proportion of the population to be covered by this tolerance interval.

side 	 Whether a 1-sided or 2-sided tolerance interval is required (determined by side = 1 or side = 2, respectively).

method 	 The method for determining which indices of the ordered observations will be used for the tolerance intervals. "WILKS" is the Wilks method, which produces tolerance bounds symmetric about the observed center of the residuals by using the beta distribution. "WALD" is the Wald method, which produces (possibly) multiple tolerance bounds for side = 2 (each having at least the specified confidence level), but is the same as method = "WILKS" for side = 1. "HM" is the Hahn-Meeker method, which is based on the binomial distribution, but the upper and lower bounds may exceed the minimum and maximum of the sample data. For side = 2, this method will yield two intervals if an odd number of observations are to be trimmed from each side.

upper 	 The upper bound of the data. When NULL, then the maximum of x is used.

lower 	 The lower bound of the data. When NULL, then the minimum of x is used.

Value

nptol.int returns a data frame with items:

alpha 	 The specified significance level.

P 	 The proportion of the population covered by this tolerance interval.

1-sided.lower 	 The 1-sided lower tolerance bound. This is given only if side = 1.

1-sided.upper 	 The 1-sided upper tolerance bound. This is given only if side = 1.

2-sided.lower 	 The 2-sided lower tolerance bound. This is given only if side = 2.

2-sided.upper 	 The 2-sided upper tolerance bound. This is given only if side = 2.

References

Bury, K. (1999), Statistical Distributions in Engineering, Cambridge University Press.

Hahn, G. J. and Meeker, W. Q. (1991), Statistical Intervals: A Guide for Practitioners, Wiley-Interscience.

Wald, A. (1943), An Extension of Wilks' Method for Setting Tolerance Limits, The Annals of Mathematical Statistics, 14, 45–55.

Wilks, S. S. (1941), Determination of Sample Sizes for Setting Tolerance Limits, The Annals of Mathematical Statistics, 12, 91–96.

See Also

distfree.est, npregtol.int

Examples

 
## 90%/95% 2-sided nonparametric tolerance intervals for a 
## sample of size 20. 

set.seed(100)
x <- rlogis(20, 5, 1)
out <- nptol.int(x = x, alpha = 0.10, P = 0.95, side = 1,
                 method = "WILKS", upper = NULL, lower = NULL)
out

plottol(out, x, plot.type = "both", side = "two", x.lab = "X")

[Package tolerance version 0.2.3 Index]