normtol.int(tolerance)	R Documentation

Normal (or Log-Normal) Tolerance Intervals

Description

Provides 1-sided or 2-sided tolerance intervals for data distributed according to either a normal distribution or log-normal distribution.

Usage


normtol.int(x, alpha = 0.05, P = 0.99, side = 1,
            method = c("HE", "WBE"), log.norm = FALSE)

Arguments

x 	 A vector of data which is distributed according to either a normal distribution or a log-normal distribution.

alpha 	 The level chosen such that 1-alpha is the confidence level.

P 	 The proportion of the population to be covered by this tolerance interval.

side 	 Whether a 1-sided or 2-sided tolerance interval is required (determined by side = 1 or side = 2, respectively).

method 	 The method for calculating the k-factors. The k-factor for the 1-sided tolerance intervals is performed exactly and thus the same for either method chosen. "HE" is the Howe method and is often viewed as being extremely accurate, even for small sample sizes. "WBE" is the Weissberg-Beatty method, which performs similarly to the Howe method for larger sample sizes. Note that the "ELL" option for the Ellison correction is disabled here.

log.norm 	 If TRUE, then the data is considered to be from a log-normal distribution, in which case the output gives tolerance intervals for the log-normal distribution. The default is FALSE.

Details

Recall that if the random variable X is distributed according to a log-normal distribution, then the random variable Y = ln(X) is distributed according to a normal distribution.

Value

normtol.int returns a data frame with items:

alpha 	 The specified significance level.

P 	 The proportion of the population covered by this tolerance interval.

x.bar 	 The sample mean.

1-sided.lower 	 The 1-sided lower tolerance bound. This is given only if side = 1.

1-sided.upper 	 The 1-sided upper tolerance bound. This is given only if side = 1.

2-sided.lower 	 The 2-sided lower tolerance bound. This is given only if side = 2.

2-sided.upper 	 The 2-sided upper tolerance bound. This is given only if side = 2.

References

Howe, W. G. (1969), Two-Sided Tolerance Limits for Normal Populations - Some Improvements, Journal of the American Statistical Association, 64, 610–620.

Weissberg, A. and Beatty, G. (1969), Tables of Tolerance Limit Factors for Normal Distributions, Technometrics, 2, 483–500.

See Also

Normal, K.factor

Examples

 
## 95%/95% 2-sided normal tolerance intervals for a sample
## of size 100. 

set.seed(100)
x <- rnorm(100, 0, 0.2)
out <- normtol.int(x = x, alpha = 0.05, P = 0.95, side = 2,
                   method = "HE", log.norm = FALSE)
out

plottol(out, x, plot.type = "both", side = "two", 
        x.lab = "Normal Data")

[Package tolerance version 0.2.3 Index]