poistol.int(tolerance)	R Documentation

Poisson Tolerance Intervals

Description

Provides 1-sided or 2-sided tolerance intervals for Poisson random variables. From a statistical quality control perspective, these limits bound the number of occurrences (which follow a Poisson distribution) in a specified future time period.

Usage


poistol.int(x, n, m, alpha = 0.05, P = 0.99, side = 1, 
            method = c("TAB", "LS"))

Arguments

x 	 The number of occurrences of the event in time period n.

n 	 The time period of the original measurements.

m 	 The specified future length of time.

alpha 	 The level chosen such that 1-alpha is the confidence level.

P 	 The proportion of occurrences in future time lengths of size m to be covered by this tolerance interval.

side 	 Whether a 1-sided or 2-sided tolerance interval is required (determined by side = 1 or side = 2, respectively).

method 	 The method for calculating the lower and upper confidence bounds, which are used in the calculation of the tolerance bounds. The default method is "TAB", which is the tabular method and is usually preferred for a smaller number of occurrences. "LS" gives the large-sample method, which is usually preferred when the number of occurrences is x>20.

Value

poistol.int returns a data frame with items:

alpha 	 The specified significance level.

P 	 The proportion of occurrences in future time periods of length m.

lambda.hat 	 The mean occurrence rate per unit time, calculated by x/n.

1-sided.lower 	 The 1-sided lower tolerance bound. This is given only if side = 1.

1-sided.upper 	 The 1-sided upper tolerance bound. This is given only if side = 1.

2-sided.lower 	 The 2-sided lower tolerance bound. This is given only if side = 2.

2-sided.upper 	 The 2-sided upper tolerance bound. This is given only if side = 2.

References

Hahn, G. J. and Chandra, R. (1981), Tolerance Intervals for Poisson and Binomial Variables, Journal of Quality Technology, 13, 100–110.

See Also

Poisson

Examples

 
## 95%/90% 1-sided Poisson tolerance limits for future 
## occurrences in a period of length of 3.  Both methods
## are presented for comparison.

poistol.int(x = 45, n = 9, m = 3, alpha = 0.05, P = 0.90,
            side = 1, method = "TAB")
poistol.int(x = 45, n = 9, m = 3, alpha = 0.05, P = 0.90,
            side = 1, method = "LS")

[Package tolerance version 0.2.3 Index]