tolerance-internal(tolerance)	R Documentation

Generation of a Random Wishart Matrix

Description

Internal function for the package tolerance.

Usage


rwishart(df, p)

Arguments

df 	 The degrees of freedom for the Wishart matrix to be generated.

p 	 The dimension of the random Wishart matrix.

Details

This function is not intended to be called by the user.

See Also

mvtol.region

[Package tolerance version 1.0 Index]