umatol.int(tolerance)	R Documentation

Uniformly Most Accurate Upper Tolerance Limits for Certain Discrete Distributions

Description

Provides uniformly most accurate upper tolerance limits for the binomial, negative binomial, and Poisson distributions.

Usage


umatol.int(x, n = NULL, dist = c("Bin", "NegBin", "Pois"), N, 
           alpha = 0.05, P = 0.99)

Arguments

x 	 A vector of data which is distributed according to one of the binomial, negative binomial, or Poisson distributions. If the length of x is 1, then it is assumed that this number is the sum of iid values from the assumed distribution.

n 	 The sample size of the data. If null, then n is calculated as the length of x. If the length of x is 1, then n must be a specified positive integer.

dist 	 The distribution for the data given by x. The options are "Bin" for the binomial distribution, "NegBin" for the negative binomial distribution, and "Pois" for the Poisson distribution.

N 	 Must be specified for the binomial and negative binomial distributions. If dist = "Bin", then N is the number of Bernoulli trials and must be a positive integer. If dist = "NegBin", then N is the total number of successful trial (or dispersion parameter) and must be strictly positive.

alpha 	 The level chosen such that 1-alpha is the confidence level.

P 	 The proportion of the population to be covered by this tolerance interval.

Value

umatol.int returns a data frame with items:

alpha 	 The specified significance level.

P 	 The proportion of the population covered by this tolerance interval.

p.hat 	 The maximum likelihood estimate for the probability of success in each trial; reported if dist = "Bin".

nu.hat 	 The maximum likelihood estimate for the probability of success in each trial; reported if dist = "NegBin".

lambda.hat 	 The maximum likelihood estimate for the rate of success; reported if dist = "Pois".

1-sided.upper 	 The 1-sided upper tolerance limit.

References

Zacks, S. (1970), Uniformly Most Accurate Tolerance Limits for Monotone Likelihood Ratio Families of Discrete Distributions, Journal of the American Statistical Association, 65, 307–316.

See Also

Binomial, NegBinomial, Poisson

Examples

 
## Examples from Zacks (1970).

umatol.int(25, n = 4, dist = "Bin", N = 10, alpha = 0.10, 
           P = 0.95)
umatol.int(13, n = 10, dist = "NegBin", N = 2, alpha = 0.10,
           P = 0.95)
umatol.int(37, n = 10, dist = "Pois", alpha = 0.10, P = 0.95)

[Package tolerance version 0.2.3 Index]